refactor: js file separation
