+++ packages/borapark-vanilla/js/constant.js
@@ -1,0 +1,28 @@
+const DRINK_NAME = {
+  COLA: "콜라",
+  SPRITE: "속이 사이다",
+  FANTA: "판타지판타",
+  EOMUK: "오뎅국물",
+  LATTE: "부장라떼",
+  WATER: "오아시스",
+  RED_BULL: "레드뿔",
+  HOT_SEVEN: "핫세븐",
+  COFFEE_MILK: "커피우유",
+  AMERICANO: "아메리카노",
+  SKY_BORI: "하늘보리",
+};
+const DRINK_PRICES = {
+  COLA: 1500,
+  SPRITE: 1700,
+  FANTA: 1500,
+  EOMUK: 1800,
+  LATTE: 800,
+  WATER: 1000,
+  RED_BULL: 2000,
+  HOT_SEVEN: 1900,
+  COFFEE_MILK: 1400,
+  AMERICANO: 2000,
+  SKY_BORI: 1500,
+};
+
+export { DRINK_NAME, DRINK_PRICES };
